Group allocation size by class & simplify freelist
Refactored the handling of superblock sizes so that the size of the
blocks and the whole superblock is determined by m_size_class. This
replaces m_block_size and maked it possible to unify block allocation
and freeing accross different size classes - this was handled by
different member functions previously.

Removed the superblock-specific freelists. There is only a single
freelist array now which is indexed by size class. This greatly
simplified the freelist and quarantine handling by removing
special cases to handle different allocation sizes.

Refactored GCNodeAllocator::allocate() so that it directly tries to
allocate from a freelist, instead of delegating to AllocatorSuperblock.
Allocation of large non-superblock allocations is now handled in
GCNodeAllocator.

Changed naming style for all allocator classes so that it better maches
the existing Rho style.